Refactored code and fixed type errors
